% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barracuda-package.R
\docType{package}
\name{barracuda-package}
\alias{barracuda}
\alias{barracuda-package}
\title{barracuda: barracuda}
\description{
A simple R package that uses Rcpp and NVIDIA's CUDA. This is a learning exercise.
}
\keyword{internal}
