% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barracuda.R
\name{cuBLASdgemm}
\alias{cuBLASdgemm}
\title{cuBLASdgemm}
\usage{
cuBLASdgemm(A, B)
}
\arguments{
\item{A}{Left matrix}

\item{B}{Right matrix}
}
\value{
The result of \code{A \%*\% B}.
}
\description{
The \code{cuBLASdgemm} function multiplies two matrices on the GPU.
}
