% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barracuda.R
\name{cuBLASdgemm_ABAt}
\alias{cuBLASdgemm_ABAt}
\title{cuBLASdgemm_ABAt}
\usage{
cuBLASdgemm_ABAt(A, B)
}
\arguments{
\item{A}{Left and right (transposed) matrix}

\item{B}{Right matrix}
}
\value{
The result of \code{A \%*\% B \%*\% t(A)}.
}
\description{
The \code{cuBLASdgemm_ABAt} function computes ABAt on the GPU.
}
