% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barracuda.R
\name{cuBLASsgemm}
\alias{cuBLASsgemm}
\title{cuBLASsgemm}
\usage{
cuBLASsgemm(A, B)
}
\arguments{
\item{A}{Left matrix}

\item{B}{Right matrix}
}
\value{
The result of \code{A \%*\% B}.
}
\description{
The \code{cuBLASsgemm} function multiplies two matrices on the GPU (float32).
}
