% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barracuda.R
\name{dgemm}
\alias{dgemm}
\title{dgemm}
\usage{
dgemm(A, B)
}
\arguments{
\item{A}{Left matrix}

\item{B}{Right matrix}
}
\value{
The result of \code{A \%*\% B}.
}
\description{
The \code{dgemm} function multiplies two matrices on the GPU.
}
