% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barracuda.R
\name{sgemm_naive}
\alias{sgemm_naive}
\title{sgemm_naive}
\usage{
sgemm_naive(A, B)
}
\arguments{
\item{A}{Left matrix}

\item{B}{Right matrix}
}
\value{
The result of \code{A \%*\% B}.
}
\description{
The \code{sgemm_naive} function multiplies two float32 matrices on the GPU using
a naive kernel.
}
